requirements.txt
pymupdf
sentence-transformers
faiss-cpu
nltk
numpy
tqdm
